refactor: split code onto modules
